requirements.txt
streamlit
edge-tts
google-generativeai